+++ Sources/StreamChatUI/Composer/ComposerVC.swift
@@ -79,6 +79,9 @@ open class ComposerVC: _ViewController,
         public var cooldownTime: Int
         /// A boolean value indicating if the message url enrichment should be skipped.
         public var skipEnrichUrl: Bool
+        /// A boolean value indicating if the message should be shown in the channel.
+        /// If the provided value is nil, it won't change the current checkbox state.
+        public var showReplyInChannel: Bool?
 
         /// A boolean that checks if the message contains any content.
         public var isEmpty: Bool {
@@ -88,6 +91,14 @@ open class ComposerVC: _ViewController,
             }
             // All other cases
             return text.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty && attachments.isEmpty
+        }
+
+        /// The text that should be sent to the backend considering the command.
+        public var inputText: String {
+            if let command = command {
+                return "/\(command.name) " + text
+            }
+            return text.trimmingCharacters(in: .whitespacesAndNewlines)
         }
 
         /// A boolean that checks if the composer is replying in a thread
@@ -123,7 +134,8 @@ open class ComposerVC: _ViewController,
             command: Command?,
             extraData: [String: RawJSON] = [:],
             cooldownTime: Int = 0,
-            skipEnrichUrl: Bool = false
+            skipEnrichUrl: Bool = false,
+            showReplyInChannel: Bool? = nil
         ) {
             self.text = text
             self.state = state
@@ -136,6 +148,7 @@ open class ComposerVC: _ViewController,
             self.extraData = extraData
             self.cooldownTime = cooldownTime
             self.skipEnrichUrl = skipEnrichUrl
+            self.showReplyInChannel = showReplyInChannel
         }
 
         /// Creates a new content struct with all empty data.
@@ -169,6 +182,26 @@ open class ComposerVC: _ViewController,
                 extraData: [:],
                 cooldownTime: cooldownTime,
                 skipEnrichUrl: false
+            )
+        }
+
+        /// Sets the content state to new message from a saved draft.
+        ///
+        /// - Parameter message: The message which was saved as a draft.
+        public mutating func draftMessage(_ message: DraftMessage) {
+            self = .init(
+                text: message.text,
+                state: message.quotedMessage != nil ? .quote : .new,
+                editingMessage: nil,
+                quotingMessage: message.quotedMessage,
+                threadMessage: threadMessage,
+                attachments: message.attachments.toAnyAttachmentPayload(),
+                mentionedUsers: message.mentionedUsers,
+                command: message.command.map { Command(name: $0, args: message.text) },
+                extraData: message.extraData,
+                cooldownTime: cooldownTime,
+                skipEnrichUrl: skipEnrichUrl,
+                showReplyInChannel: message.showReplyInChannel
             )
         }
 
@@ -484,6 +517,12 @@ open class ComposerVC: _ViewController,
         composerView.pin(to: view)
     }
 
+    override open func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+
+        updateDraftMessageIfNeeded()
+    }
+
     open func setupAttachmentsView() {
         addChildViewController(attachmentsVC, embedIn: composerView.inputMessageView.attachmentsViewContainer)
         attachmentsVC.didTapRemoveItemButton = { [weak self] index in
@@ -684,6 +723,10 @@ open class ComposerVC: _ViewController,
             } else {
                 composerView.checkboxControl.label.text = L10n.Composer.Checkmark.channelReply
             }
+
+            if let showReplyInChannel = content.showReplyInChannel {
+                composerView.checkboxControl.isSelected = showReplyInChannel
+            }
         }
     }
 
@@ -773,12 +816,9 @@ open class ComposerVC: _ViewController,
             return
         }
 
-        let text: String
-        if let command = content.command {
-            text = "/\(command.name) " + content.text
-        } else {
-            text = content.text.trimmingCharacters(in: .whitespacesAndNewlines)
-        }
+        deleteDraftMessageIfNeeded()
+
+        let text = content.inputText
 
         if let editingMessage = content.editingMessage {
             editMessage(withId: editingMessage.id, newText: text)
@@ -913,14 +953,7 @@ open class ComposerVC: _ViewController,
     open func createNewMessage(text: String) {
         guard let cid = channelController?.cid else { return }
 
-        // If the user included some mentions via suggestions,
-        // but then removed them from text, we should remove them from
-        // the content we'll send
-        for user in content.mentionedUsers {
-            if !text.contains(mentionText(for: user)) {
-                content.mentionedUsers.remove(user)
-            }
-        }
+        removeMentionUserIfNotIncluded(in: text)
 
         if let threadParentMessageId = content.threadMessage?.id {
             let messageController = channelController?.client.messageController(
@@ -950,6 +983,67 @@ open class ComposerVC: _ViewController,
             skipEnrichUrl: content.skipEnrichUrl,
             extraData: content.extraData
         )
+    }
+
+    /// Updates the draft message in the channel or thread if draft messages are enabled.
+    ///
+    /// The draft is created from the current content in the composer.
+    private func updateDraftMessageIfNeeded() {
+        guard components.isDraftMessagesEnabled else {
+            return
+        }
+
+        let text = content.command != nil ? content.text : content.inputText
+        if content.isEmpty {
+            return
+        }
+
+        removeMentionUserIfNotIncluded(in: text)
+
+        if let threadParentMessageId = content.threadMessage?.id, let cid = channelController?.cid {
+            let messageController = channelController?.client.messageController(
+                cid: cid,
+                messageId: threadParentMessageId
+            )
+
+            messageController?.updateDraftReply(
+                text: text,
+                attachments: content.attachments,
+                mentionedUserIds: content.mentionedUsers.map(\.id),
+                quotedMessageId: content.quotingMessage?.id,
+                showReplyInChannel: composerView.checkboxControl.isSelected,
+                command: content.command,
+                extraData: content.extraData
+            )
+            return
+        }
+
+        channelController?.updateDraftMessage(
+            text: text,
+            attachments: content.attachments,
+            mentionedUserIds: content.mentionedUsers.map(\.id),
+            quotedMessageId: content.quotingMessage?.id,
+            command: content.command,
+            extraData: content.extraData
+        )
+    }
+
+    /// Deletes the draft message in the channel or thread if there is one.
+    private func deleteDraftMessageIfNeeded() {
+        if let threadParentMessageId = content.threadMessage?.id, let cid = channelController?.cid {
+            let messageController = channelController?.client.messageController(
+                cid: cid,
+                messageId: threadParentMessageId
+            )
+            if messageController?.message?.draftReply != nil {
+                messageController?.deleteDraftReply()
+            }
+            return
+        }
+
+        if channelController?.channel?.draftMessage != nil {
+            channelController?.deleteDraftMessage()
+        }
     }
 
     /// Updates an existing message.
@@ -1450,6 +1544,11 @@ open class ComposerVC: _ViewController,
         guard textView.text != content.text else { return }
 
         content.text = textView.text
+
+        /// If the input message was erased and there is a draft message, the draft message should be deleted.
+        if content.isEmpty {
+            deleteDraftMessageIfNeeded()
+        }
     }
 
     open func textView(
@@ -1662,6 +1761,17 @@ open class ComposerVC: _ViewController,
         actions.forEach(alert.addAction)
 
         present(alert, animated: true)
+    }
+
+    private func removeMentionUserIfNotIncluded(in currentText: String) {
+        // If the user included some mentions via suggestions,
+        // but then removed them from text, we should remove them from
+        // the content we'll send
+        for user in content.mentionedUsers {
+            if !currentText.contains(mentionText(for: user)) {
+                content.mentionedUsers.remove(user)
+            }
+        }
     }
 }