+++ Tests/StreamChatTests/APIClient/Endpoints/Payloads/ChannelListPayload_Tests.swift
@@ -247,7 +247,8 @@ final class ChannelListPayload_Tests: XCTestCase {
                         unreadMessagesCount: (0..<10).randomElement()!
                     )
                 },
-                isHidden: false
+                isHidden: false,
+                draft: nil
             )
         }