+++ CHANGELOG.md
@@ -6,9 +6,30 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## StreamChat
 ### ✅ Added
 - Add sending messages to only specific members [#3595](https://github.com/GetStream/stream-chat-swift/pull/3595)
+- Add Draft Messages Support [#3588](https://github.com/GetStream/stream-chat-swift/pull/3588)
+  - Add `ChatChannel.draftMessage`
+  - Add `ChatMessage.draftReply`
+  - Add `ChannelController`:
+     - `updateDraftMessage()`
+     - `deleteDraftMessage()`
+     - `loadDraftMessage()`
+  - Add `MessageController`:
+     - `updateDraftReply()`
+     - `deleteDraftReply()`
+     - `loadDraftReply()`
+  - Add `CurrentUserController`:
+     - `deleteDraft()`
+     - `loadDraftMessages()`
+     - `loadMoreDraftMessages()` 
+
 ### 🐞 Fixed
 - Update channel's preview message when coming back to online [#3574](https://github.com/GetStream/stream-chat-swift/pull/3574)
-### 🔄 Changed
+- Fix message transformer not being applied when editing a message [#3602](https://github.com/GetStream/stream-chat-swift/pull/3602)
+
+## StreamChatUI
+### ✅ Added
+- Add `Components.isDraftMessagesEnabled` to enable Draft Messages [#3588](https://github.com/GetStream/stream-chat-swift/pull/3588)
+- Add draft preview in Channel List and Thread List if drafts are enabled [#3588](https://github.com/GetStream/stream-chat-swift/pull/3588)
 
 # [4.72.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.72.0)
 _February 04, 2025_